sdk: split query_verbs into module directory (8/9: extract hotspots)
Move hotspots and helpers to query_verbs/hotspots.rs.
